pyproject.toml
[tool.poetry]
name = "invoice-manager"
version = "0.1.0"
description = ""
authors = ["Magnus Glasder <mglasder@ethz.ch>"]
readme = "README.md"

[tool.poetry.scripts]
invoice = "invoice_manager.app:app"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.0"
typer = "^0.4.0"
black = "^22.1.0"
prettytable = "^3.0.0"
pandas = "^1.4.0"
google-api-python-client = "^2.45.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
PyYAML = "^6.0"
yagmail = "^0.15.277"
orjson = "^3.8.9"
pydantic-yaml = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
